analysis/requirements.txt
clickhouse-connect>=0.7.0
psycopg2-binary>=2.9.9
redis>=5.0.0
numpy>=1.26.0
scipy>=1.11.0
python-dotenv>=1.0.0

